+++ lib/src/discrete_mathematics/linear_algebra/vector/base/vector_base.dart
@@ -1,5 +1,7 @@
 import 'dart:math';
 
+import 'package:quiver/core.dart';
+
 import '../../exceptions/vector_exception.dart';
 import '../../matrix/base/matrix_base.dart';
 
@@ -12,30 +14,30 @@ abstract class VectorBase {
   VectorBase.init(this.data);
 
   /// Data for vector
-  List<double> data;
+  List<num> data;
 
   /// Count of vector's numbers
   int get itemCount => data.length;
 
   /// Gets length of this vector
-  double get length => euclideanNorm();
+  num get length => euclideanNorm();
 
   /// Gets number at specified [index]
   ///
   /// [index] is in range from 1 to end inclusively.
-  double itemAt(int index) => data[index - 1];
+  num itemAt(int index) => data[index - 1];
 
   /// Sets [item] in specified [position] of vector
   ///
   /// [position] should starts from 1.
-  void setItem(int position, double item) {
+  void setItem(int position, num item) {
     data[position - 1] = item;
   }
 
   /// Gets norm of vector alse known as vector's length
   ///
   /// [p] should have only integer value, if not - any fractional digits will be discarded.
-  double norm(double p) {
+  num norm(double p) {
     final n = p.truncateToDouble();
     if (n <= 0) {
       throw VectorException('P must be greater or equal to 1! Given $p.');
@@ -57,13 +59,13 @@ abstract class VectorBase {
   }
 
   /// Gets Euclidean norm
-  double euclideanNorm() => norm(2);
+  num euclideanNorm() => norm(2);
 
   /// Gets the norm where p is infinite
-  double maxNorm() => norm(double.infinity);
+  num maxNorm() => norm(double.infinity);
 
   /// Multiply this vector by [vector] using dot product algorithm
-  double dotProduct(covariant VectorBase vector) =>
+  num dotProduct(covariant VectorBase vector) =>
       toMatrix().transpose().multiplyByMatrix(vector.toMatrix()).itemAt(1, 1);
 
   /// Gets cross product of this vector and another [vector]
@@ -79,7 +81,7 @@ abstract class VectorBase {
   ///
   /// Dafault unit for measuring angle is `radian`.
   /// If [degrees] is true - result will have the degrees unit.
-  double angleBetween(covariant VectorBase vector, {bool degrees = false}) {
+  num angleBetween(covariant VectorBase vector, {bool degrees = false}) {
     final dot = dotProduct(vector);
     final magnitudes = euclideanNorm() * vector.euclideanNorm();
     if (degrees == false) {
@@ -95,7 +97,7 @@ abstract class VectorBase {
   VectorBase hadamardProduct(covariant VectorBase vector);
 
   /// Transform each element of this vector and return transformed vector
-  VectorBase transform(double t(double value));
+  VectorBase transform(num t(num value));
 
   /// Checks if vector is unit vector
   bool isUnit() => euclideanNorm() == 1;
@@ -112,11 +114,19 @@ abstract class VectorBase {
 
   /// Divides corresponding values of this vactors by [vector]
   VectorBase operator /(covariant VectorBase vector) =>
-    this * vector.transform((v) => 1 / v);
+      this * vector.transform((v) => 1 / v);
 
   /// Add values of this vector to [vector]'s values
   VectorBase operator +(covariant VectorBase vector);
 
   /// Subtract [vector]'s values from values of this vector
-  VectorBase operator -(covariant VectorBase vector) => this + vector.transform((v) => -v);
+  VectorBase operator -(covariant VectorBase vector) =>
+      this + vector.transform((v) => -v);
+
+  @override
+  bool operator ==(Object other) =>
+      other is VectorBase && hashCode == other.hashCode;
+
+  @override
+  int get hashCode => hashObjects(data);
 }