+++ lib/src/general/elementary_algebra/quadratic_equation.dart
@@ -8,34 +8,32 @@ class QuadraticEquation extends EquationBase {
   /// Accepts coefficients [a], [b] and [c]
   ///
   /// [a] value shouldn't be equal to zero. Otherwise the [EquationException] is thrown.
-  QuadraticEquation(double a, {this.b, this.c}) {
+  QuadraticEquation({this.a = 1, this.b = 0, this.c = 0}) {
     if (a == 0) {
       throw EquationException('Coefficient a shouldn\'t be equal to zero!');
-    } else {
-      this.a = a;
     }
   }
 
   /// [a] coefficient in quadratic equation
-  double a;
+  num a;
 
   /// [b] coefficient in quadratic equation
-  double b = 0;
+  num b;
 
   /// [c] coefficient in quadratic equation
-  double c = 0;
+  num c;
 
   @override
-  Set<double> calculate() {
-    final result = Set<double>();
+  Set<num> calculate() {
+    final result = Set<num>();
     final dis = discriminant();
 
     if (dis > 0) {
       for (var i = 1; i <= 2; i++) {
-        result.add((-b + pow(-1, i) * sqrt(dis)) / 2 * a);
+        result.add((-b + pow(-1, i) * sqrt(dis)) / (2 * a));
       }
     } else if (dis == 0) {
-      result.add(-b / 2 * a);
+      result.add(-b / (2 * a));
     } else {
       throw EquationException(
           'Expression haven\'t solutions, because of discriminant is less than zero!');
@@ -44,10 +42,10 @@ class QuadraticEquation extends EquationBase {
   }
 
   @override
-  double discriminant() => pow(b, 2) - 4 * a * c;
+  num discriminant() => pow(b, 2) - 4 * a * c;
 
   @override
-  bool evaluateForZero(double x) => (a * pow(x, 2) + b * x + c) == 0;
+  bool evaluateForZero(num x) => (a * pow(x, 2) + b * x + c) == 0;
 
   @override
   String toString() => '$a * x2 + $b * x + $c';